+++ CHANGELOG.md
@@ -46,6 +46,7 @@ script to proceed.
 - Avoid modal for dashboard refresh error.
 - temBoard agent auto_configure.sh conditionnaly enable statements plugins.
 - temBoard agent auto_configure.sh now generates a single configuration file.
+- Dropped temBoard agent HTTP endpoint /monitoring/probe/*.
 
 
 ## [7.11] - Unreleased